refactoring controller complete

+++ controller.js
@@ -39,14 +39,17 @@ var FEATURES = {
 		var subscribers;
 
 		step(function() {
-			 var g = this.group();
-			 for(var id in req.items) {
-			     if (req.items.hasOwnProperty(id))
-				 t.writeItem(req.from, req.node, id, req.items[id], g());
+			 if (!req.items)
+			     this(null, []);
+			 else {
+			     var g = this.group();
+			     for(var id in req.items) {
+				 if (req.items.hasOwnProperty(id))
+				     t.writeItem(req.from, req.node, id, req.items[id], g());
+			     }
 			 }
 		     }, cb);
 	    },
-	    /* TODO: */
 	    subscriberNotification: function(req, subscribers) {
 		subscribers.forEach(function(subscriber) {
 		    callFrontend('notify', subscriber, req.node, req.items);
@@ -61,13 +64,16 @@ var FEATURES = {
 
 		/* TODO: check perms */
 		step(function() {
-			 var g = this.group();
-			 req.itemIds.forEach(function(itemId) {
-			     t.deleteItem(req.node, itemId, g());
-			 });
+			 if (req.itemIds.length < 1)
+			     this(null, []);
+			 else {
+			     var g = this.group();
+			     req.itemIds.forEach(function(itemId) {
+				 t.deleteItem(req.node, itemId, g());
+			     });
+			 }
 		     }, cb);
 	    },
-	    /* TODO: */
 	    subscriberNotification: function(req, subscribers) {
 		subscribers.forEach(function(subscriber) {
 		    callFrontend('retracted', subscriber, req.node, req.itemIds);
@@ -84,11 +90,16 @@ var FEATURES = {
 		     }, function(err, ids_) {
 			 if (err) throw err;
 
+console.log("item ids: " + JSON.stringify(ids_));
 			 ids = ids_;
-			 var g = this.group();
-			 ids.forEach(function(id) {
-					 t.getItem(req.node, id, g());
-				     });
+			 if (ids.length < 1)
+			     this(null, []);
+			 else {
+			     var g = this.group();
+			     ids.forEach(function(id) {
+				 t.getItem(req.node, id, g());
+			     });
+			 }
 		     }, function(err, items) {
 			 if (err) throw err;
 
@@ -135,6 +146,9 @@ exports.request = function(req) {
 	req.callback(new Error('not-implemented'));
 	return;
     }
+    var debug = function(s) {
+	console.log(req.from + ' >> ' + req.feature + '/' + req.operation + ': ' + s);
+    };
 
     if (operation.needOwner) {
 	/* If ownership were not hard-coded anymore this had to be
@@ -143,7 +157,7 @@ exports.request = function(req) {
 	var nodeM = req.node.match(/^\/user\/(.+?)\/([a-zA-Z0-9\/_\-]+)$/);
 	var userM = req.from.match(/^(.+?):(.+)$/);
 	if (!nodeM || nodeM[1] !== userM[2]) {
-	    cb(new Error('forbidden'));
+	    req.callback(new Error('forbidden'));
 	    return;
 	}
     }
@@ -154,20 +168,76 @@ exports.request = function(req) {
 	    return;
 	}
 
-	operation.transaction(req, t, function(err) {
-	    if (err)
-		t.rollback(requestFinalizer(req, arguments));
-	    else
-		t.commit(requestFinalizer(req, arguments));
+	var transactionResults;
+	/* Run operation transaction first */
+	var steps = [function() {
+			 debug('transaction');
+			 operation.transaction(req, t, this);
+		     }, function(err) {
+			 debug('transaction done');
+			 if (err) throw err;
+
+			 /* Regardless of the following steps, we pass
+			  * the operation's transaction result to the
+			  * final callback.
+			  */
+			 transactionResults = arguments;
+			 /* And continue:
+			  */
+			 this(null);
+		     }];
+        var subscribers;
+	if (operation.subscriberNotification) {
+	    /* For subscriber notification, get the list of subscribers
+	     * while still inside transaction.
+	     */
+	    steps.push(function(err) {
+		if (err) throw err;
+
+		t.getSubscribers(req.node, this);
+	    }, function(err, subscribers_) {
+		if (err) throw err;
+
+		subscribers = subscribers_;
+		this(null);
+	    });
+	}
+	/* Finalize transaction
+	 */
+	steps.push(function(err) {
+	    if (err) {
+		var that = this;
+		debug('transaction rollback');
+		t.rollback(function() {
+		    /* Keep error despite successful rollback */
+		    that(err);
+		});
+	    } else {
+		debug('transaction commit');
+		t.commit(this);
+	    }
 	});
+	if (operation.subscriberNotification) {
+	    /* Transaction successful? Call subscriberNotification. */
+	    steps.push(function(err) {
+		if (err) throw err;
+
+		operation.subscriberNotification(req, subscribers);
+		this(null);
+	    });
+	}
+	/* Last step: return to caller (view) */
+	steps.push(function(err) {
+	    debug('callback');
+	    if (err)
+		req.callback(err);
+	    else
+		req.callback.apply(req, transactionResults);
+	});
+
+	step.apply(null, steps);
     });
 };
-
-function requestFinalizer(req, args) {
-    return function() {
-	req.callback.apply(req, args);
-    };
-}
 
 var frontends = {};
 /**
